feat: add tests
I know I wrote code first, then test and it's wrong,
 because it was a simple case

+++ src/product/product.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreateProductDto } from './dto/create-product.dto';
@@ -16,26 +16,39 @@ export class ProductService {
   }
 
   async findOne(id: string): Promise<Product> {
-    return await this.model.findById(id).exec();
+    const exists = await this.model.findById(id).exec();
+    if (!exists) {
+      throw new NotFoundException('product not found');
+    }
+    return exists;
   }
 
   async create(createProductDto: CreateProductDto): Promise<Product> {
-    return await new this.model({
+    return await this.model.create({
       ...createProductDto,
       createdAt: new Date(),
-    }).save();
+    });
   }
 
   async update(
     id: string,
     updateProductDto: UpdateProductDto,
   ): Promise<Product> {
+    const exists = await this.model.findById(id).exec();
+    if (!exists) {
+      throw new NotFoundException('product not found');
+    }
+
     return await this.model
       .findByIdAndUpdate(id, updateProductDto, { new: true })
       .exec();
   }
 
   async remove(id: string): Promise<Product> {
+    const exists = await this.model.findById(id).exec();
+    if (!exists) {
+      throw new NotFoundException('product not found');
+    }
     return await this.model.findByIdAndDelete(id).exec();
   }
 }
